cleanup: address feedback on write index validation
